# docs/swagger.yaml
basePath: /api/v1
definitions:
  internal_httpapi.authResponse:
    properties:
      token:
        type: string
      user:
        $ref: '#/definitions/task_internal_model.User'
    type: object
  internal_httpapi.commentResponse:
    properties:
      comment:
        $ref: '#/definitions/task_internal_model.Comment'
    type: object
  internal_httpapi.commentsResponse:
    properties:
      comments:
        items:
          $ref: '#/definitions/task_internal_model.Comment'
        type: array
    type: object
  internal_httpapi.createCommentRequest:
    properties:
      content:
        type: string
    type: object
  internal_httpapi.createTaskRequest:
    properties:
      assignee_id:
        type: integer
      description:
        type: string
      title:
        type: string
    type: object
...
  internal_httpapi.errorInfo:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
  internal_httpapi.errorResponse:
    properties:
      error:
        $ref: '#/definitions/internal_httpapi.errorInfo'
    type: object
  internal_httpapi.historyResponse:
    properties:
      history:
        items:
          $ref: '#/definitions/task_internal_model.TaskHistory'
        type: array
    type: object
...
  internal_httpapi.loginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
...
  internal_httpapi.registerRequest:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        type: string
    type: object
  internal_httpapi.statsResponse:
    properties:
      stats:
        $ref: '#/definitions/task_internal_model.TeamStats'
    type: object
...
  internal_httpapi.taskListResponse:
    properties:
      limit:
        type: integer
      offset:
        type: integer
      tasks:
        items:
          $ref: '#/definitions/task_internal_model.Task'
        type: array
      total:
        type: integer
    type: object
  internal_httpapi.taskResponse:
    properties:
      task:
        $ref: '#/definitions/task_internal_model.Task'
    type: object
...
  internal_httpapi.updateTaskRequest:
    properties:
      assignee_id:
        type: integer
      description:
        type: string
      status:
        $ref: '#/definitions/task_internal_model.TaskStatus'
      title:
        type: string
      version:
        type: integer
    type: object
  internal_httpapi.userResponse:
    properties:
      user:
        $ref: '#/definitions/task_internal_model.User'
    type: object
  task_internal_model.AssigneeStat:
    properties:
      closed_count:
        type: integer
      name:
        type: string
    type: object
  task_internal_model.Change:
    properties:
      field:
        type: string
      new_value: {}
      old_value: {}
    type: object
  task_internal_model.Comment:
    properties:
      content:
        type: string
      created_at:
        type: string
      id:
        type: integer
      task_id:
        type: integer
      user_id:
        type: integer
    type: object
...
  task_internal_model.Task:
    properties:
      assignee_id:
        type: integer
      closed_at:
        type: string
      created_at:
        type: string
      created_by:
        type: integer
      description:
        type: string
      id:
        type: integer
      status:
        $ref: '#/definitions/task_internal_model.TaskStatus'
      team_id:
        type: integer
      title:
        type: string
      updated_at:
        type: string
      version:
        type: integer
    type: object
  task_internal_model.TaskHistory:
    properties:
      changed_by:
        type: integer
      changes:
        items:
          $ref: '#/definitions/task_internal_model.Change'
        type: array
      created_at:
        type: string
      id:
        type: integer
      task_id:
        type: integer
    type: object
  task_internal_model.TaskStatus:
    enum:
    - todo
    - in_progress
    - done
    type: string
    x-enum-varnames:
    - StatusTodo
    - StatusInProgress
    - StatusDone
...
  task_internal_model.TeamStats:
    properties:
      avg_close_seconds:
        type: number
      comments_count:
        type: integer
      done_count:
        type: integer
      in_progress_count:
        type: integer
      todo_count:
        type: integer
      top_assignees:
        items:
          $ref: '#/definitions/task_internal_model.AssigneeStat'
        type: array
      top_assignees_done_30d:
        type: integer
    type: object
  task_internal_model.User:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: integer
      name:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: 'REST API для управления задачами внутри команд: ролевая модель, история
    изменений, кеширование списков задач и SQL-отчёт.'
  title: Task Manager API
  version: "1.0"
paths:
  /login:
    post:
      consumes:
      - application/json
      parameters:
      - description: Login payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_httpapi.loginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_httpapi.authResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_httpapi.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_httpapi.errorResponse'
      summary: Login and get a JWT
      tags:
      - auth
  /register:
    post:
      consumes:
      - application/json
      parameters:
      - description: Registration payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_httpapi.registerRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal_httpapi.userResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_httpapi.errorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/internal_httpapi.errorResponse'
      summary: Register a new user
      tags:
      - auth
  /tasks:
    get:
      parameters:
      - description: Team ID
        in: query
        name: team_id
        required: true
        type: integer
      - description: 'Filter by status: todo, in_progress, done'
        in: query
        name: status
        type: string
      - description: Filter by assignee
        in: query
        name: assignee_id
        type: integer
      - description: Page size (1..100, default 20)
        in: query
        name: limit
        type: integer
      - description: Offset (default 0)
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_httpapi.taskListResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_httpapi.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_httpapi.errorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_httpapi.errorResponse'
      security:
      - BearerAuth: []
      summary: List tasks
      tags:
      - tasks
    post:
      consumes:
      - application/json
      parameters:
      - description: Team ID
        in: query
        name: team_id
        required: true
        type: integer
      - description: Task payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_httpapi.createTaskRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal_httpapi.taskResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_httpapi.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_httpapi.errorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_httpapi.errorResponse'
      security:
      - BearerAuth: []
      summary: Create a task
      tags:
      - tasks
  /tasks/{id}:
    put:
      consumes:
      - application/json
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: integer
      - description: Update payload (only provided fields change; version is required)
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_httpapi.updateTaskRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_httpapi.taskResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_httpapi.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_httpapi.errorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_httpapi.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_httpapi.errorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/internal_httpapi.errorResponse'
      security:
      - BearerAuth: []
      summary: Update a task
      tags:
      - tasks
  /tasks/{id}/comments:
    get:
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_httpapi.commentsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_httpapi.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_httpapi.errorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_httpapi.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_httpapi.errorResponse'
      security:
      - BearerAuth: []
      summary: List task comments
      tags:
      - comments
    post:
      consumes:
      - application/json
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: integer
      - description: Comment payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_httpapi.createCommentRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal_httpapi.commentResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_httpapi.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_httpapi.errorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_httpapi.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_httpapi.errorResponse'
      security:
      - BearerAuth: []
      summary: Add a comment
      tags:
      - comments
  /tasks/{id}/history:
    get:
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_httpapi.historyResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_httpapi.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_httpapi.errorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_httpapi.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_httpapi.errorResponse'
      security:
      - BearerAuth: []
      summary: Get task history
      tags:
      - tasks
...
  /teams/{team_id}/stats:
    get:
      parameters:
      - description: Team ID
        in: path
        name: team_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_httpapi.statsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_httpapi.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_httpapi.errorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_httpapi.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_httpapi.errorResponse'
      security:
      - BearerAuth: []
      summary: Team statistics report
      tags:
      - stats
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"