# internal/sqlc/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.31.1

package sqlc

import (
	"database/sql"
	"database/sql/driver"
	"fmt"
	"time"
)

type TasksStatus string

const (
	TasksStatusTodo       TasksStatus = "todo"
	TasksStatusInProgress TasksStatus = "in_progress"
	TasksStatusDone       TasksStatus = "done"
)

func (e *TasksStatus) Scan(src interface{}) error {
	switch s := src.(type) {
	case []byte:
		*e = TasksStatus(s)
	case string:
		*e = TasksStatus(s)
	default:
		return fmt.Errorf("unsupported scan type for TasksStatus: %T", src)
	}
	return nil
}

type NullTasksStatus struct {
	TasksStatus TasksStatus `json:"tasks_status"`
	Valid       bool        `json:"valid"` // Valid is true if TasksStatus is not NULL
}

// Scan implements the Scanner interface.
func (ns *NullTasksStatus) Scan(value interface{}) error {
	if value == nil {
		ns.TasksStatus, ns.Valid = "", false
		return nil
	}
	ns.Valid = true
	return ns.TasksStatus.Scan(value)
}

// Value implements the driver Valuer interface.
func (ns NullTasksStatus) Value() (driver.Value, error) {
	if !ns.Valid {
		return nil, nil
	}
	return string(ns.TasksStatus), nil
}

type TeamMembersRole string

const (
	TeamMembersRoleOwner  TeamMembersRole = "owner"
	TeamMembersRoleAdmin  TeamMembersRole = "admin"
	TeamMembersRoleMember TeamMembersRole = "member"
)

func (e *TeamMembersRole) Scan(src interface{}) error {
	switch s := src.(type) {
	case []byte:
		*e = TeamMembersRole(s)
	case string:
		*e = TeamMembersRole(s)
	default:
		return fmt.Errorf("unsupported scan type for TeamMembersRole: %T", src)
	}
	return nil
}

type NullTeamMembersRole struct {
	TeamMembersRole TeamMembersRole `json:"team_members_role"`
	Valid           bool            `json:"valid"` // Valid is true if TeamMembersRole is not NULL
}

// Scan implements the Scanner interface.
func (ns *NullTeamMembersRole) Scan(value interface{}) error {
	if value == nil {
		ns.TeamMembersRole, ns.Valid = "", false
		return nil
	}
	ns.Valid = true
	return ns.TeamMembersRole.Scan(value)
}

// Value implements the driver Valuer interface.
func (ns NullTeamMembersRole) Value() (driver.Value, error) {
	if !ns.Valid {
		return nil, nil
	}
	return string(ns.TeamMembersRole), nil
}

type Task struct {
	ID          uint64        `json:"id"`
	TeamID      uint64        `json:"team_id"`
	Title       string        `json:"title"`
	Description string        `json:"description"`
	Status      TasksStatus   `json:"status"`
	CreatedBy   uint64        `json:"created_by"`
	AssigneeID  sql.NullInt64 `json:"assignee_id"`
	CreatedAt   time.Time     `json:"created_at"`
	UpdatedAt   time.Time     `json:"updated_at"`
	ClosedAt    sql.NullTime  `json:"closed_at"`
	Version     uint32        `json:"version"`
}

type TaskComment struct {
	ID        uint64    `json:"id"`
	TaskID    uint64    `json:"task_id"`
	UserID    uint64    `json:"user_id"`
	Content   string    `json:"content"`
	CreatedAt time.Time `json:"created_at"`
}

type TaskHistory struct {
	ID        uint64    `json:"id"`
	TaskID    uint64    `json:"task_id"`
	ChangedBy uint64    `json:"changed_by"`
	Changes   string    `json:"changes"`
	CreatedAt time.Time `json:"created_at"`
}

type Team struct {
	ID        uint64    `json:"id"`
	Name      string    `json:"name"`
	CreatedBy uint64    `json:"created_by"`
	CreatedAt time.Time `json:"created_at"`
}

type TeamMember struct {
	TeamID    uint64          `json:"team_id"`
	UserID    uint64          `json:"user_id"`
	Role      TeamMembersRole `json:"role"`
	CreatedAt time.Time       `json:"created_at"`
}

type User struct {
	ID           uint64    `json:"id"`
	Email        string    `json:"email"`
	PasswordHash string    `json:"password_hash"`
	Name         string    `json:"name"`
	CreatedAt    time.Time `json:"created_at"`
}